Moved classes to more appropriate package.
