+++ modules/core/src/com/haulmont/cuba/core/sys/persistence/EclipseLinkSessionEventListener.java
@@ -28,8 +28,8 @@ import com.haulmont.cuba.core.global.PersistenceHelper;
 import com.haulmont.cuba.core.sys.AppContext;
 import com.haulmont.cuba.core.sys.CubaEnhanced;
 import com.haulmont.cuba.core.sys.CubaEnhancingDisabled;
-import com.haulmont.cuba.core.sys.persistence.mapping.processors.MappingProcessor;
-import com.haulmont.cuba.core.sys.persistence.mapping.processors.MappingProcessorContext;
+import com.haulmont.cuba.core.sys.persistence.mapping.MappingProcessor;
+import com.haulmont.cuba.core.sys.persistence.mapping.MappingProcessorContext;
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.BooleanUtils;
 import org.eclipse.persistence.annotations.CacheCoordinationType;
