+++ modules/core/src/com/haulmont/cuba/core/sys/persistence/mapping/JoinCriteriaMappingProcessor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.haulmont.cuba.core.sys.persistence.mapping.processors;
+package com.haulmont.cuba.core.sys.persistence.mapping;
 
 import com.haulmont.cuba.core.global.AppBeans;
 import org.eclipse.persistence.expressions.Expression;
