+++ modules/core/src/com/haulmont/cuba/core/sys/persistence/mapping/SoftDeleteMappingProcessor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.haulmont.cuba.core.sys.persistence.mapping.processors;
+package com.haulmont.cuba.core.sys.persistence.mapping;
 
 import com.google.common.base.Strings;
 import com.haulmont.cuba.core.entity.SoftDelete;
